refactoring to improve code quality
