+++ app/Containers/Payment/Traits/AccountableTrait.php
@@ -14,26 +14,4 @@ trait AccountableTrait
     public function paymentAccount() {
         return $this->morphOne(PaymentAccount::class, 'accountable');
     }
-
-    public function checkIfPaymentDataIsSet(array $fields) {
-
-        foreach ($fields as $field) {
-            if ($this->getAttributeValue($field) === null) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    public function getDetailAttributes() {
-        $attributes = $this->toArray();
-
-        unset($attributes['id']);
-        unset($attributes['created_at']);
-        unset($attributes['updated_at']);
-        unset($attributes['deleted_at']);
-
-        return $attributes;
-    }
 }
